% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-expr.R
\docType{methods}
\name{expression.graph}
\alias{expression.graph}
\alias{expression.graph,function-method}
\title{Construct an expression graph}
\usage{
expression.graph(f)

\S4method{expression.graph}{`function`}(f)
}
\arguments{
\item{f}{a function for which the expression graph should be constructed}
}
\value{
returns a list with the root node, variable nodes and leaf nodes
}
\description{
Construct an expression graph from a function
}
\examples{
f <- function(x, y) {
  x + y
}
expression.graph(f)
}
