% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-grad.R
\docType{methods}
\name{grad}
\alias{grad}
\alias{grad,function-method}
\title{Compute gradients of a function}
\usage{
grad(f, argnum = 1L)

\S4method{grad}{`function`}(f, argnum = 1L)
}
\arguments{
\item{f}{a function for which gradients should be computed}

\item{argnum}{the number of the argument of the function for which gradients
are computed}
}
\value{
the gradient of f w.r.t argnum
}
\description{
Compute gradients of a function with respect to one of its
 arguments
}
\examples{
 f <- function(x, y) {
   x + y
 }
 grad(f)(1.0, 1.0)

}
