% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jar-package.R
\docType{package}
\name{jar-package}
\alias{jar-package}
\alias{adjs}
\title{jar}
\format{
An object of class \code{list} of length 8.
}
\usage{
adjs
}
\description{
Implementation of reverse-mode autodiff primivites to take gradients of
arbitrary differentiable functions
}
\author{
Simon Dirmeier
}
\keyword{package}
